requirements.txt
a2a-sdk==v1.0.0
uvicorn
typing_extensions
sse-starlette
fastapi
jsonschema
google-adk==1.27.0